Dictionary of example data

This dataset is created from a noramalized set of tables to allow developers to add or change example datasets in a very compact and consistent way.

Usage

example_dictionary(remove_id = TRUE)

Arguments

remove_id

Remove table id's?

Value

A tibble().

Examples

example_dictionary() |> print(n = Inf)
#> # A tibble: 30 × 4
#>    aka           column                           table     value 
#>    <chr>         <chr>                            <chr>     <chr> 
#>  1 id            companies_id                     companies a     
#>  2 cluster       clustered                        companies a     
#>  3 uid           activity_uuid_product_uuid       companies a     
#>  4 xsector       sector                           companies total 
#>  5 xsubsector    subsector                        companies energy
#>  6 tsector       tilt_sector                      companies a     
#>  7 tsubsector    tilt_subsector                   companies a     
#>  8 tsubsector    tilt_subsector                   products  a     
#>  9 scenario_type type                             companies ipr   
#> 10 xsector       sector                           scenarios total 
#> 11 xsubsector    subsector                        scenarios energy
#> 12 xyear         year                             scenarios 2050  
#> 13 co2reduce     reductions                       scenarios 1     
#> 14 scenario_type type                             scenarios ipr   
#> 15 scenario_name scenario                         scenarios a     
#> 16 uid           activity_uuid_product_uuid       products  a     
#> 17 tsector       tilt_sector                      products  a     
#> 18 xunit         unit                             products  a     
#> 19 isic          isic_4digit                      products  '1234'
#> 20 co2footprint  co2_footprint                    products  1     
#> 21 uid           activity_uuid_product_uuid       inputs    a     
#> 22 iuid          input_activity_uuid_product_uuid inputs    a     
#> 23 itsector      input_tilt_sector                inputs    a     
#> 24 itsubsector   input_tilt_subsector             inputs    a     
#> 25 iunit         input_unit                       inputs    a     
#> 26 iisic         input_isic_4digit                inputs    '1234'
#> 27 ico2footprint input_co2_footprint              inputs    1     
#> 28 scenario_type type                             inputs    ipr   
#> 29 xsector       sector                           inputs    total 
#> 30 xsubsector    subsector                        inputs    energy