Calculate the ICTR indicator

The Input Product Carbon Transition Risk Indicator assesses the transition risk of the input products due to their relative carbon footprint to other input products. As a default option, each input product is compared to the carbon footprint of every other input product. Alternatively, users can also choose to compare those input products of the same type, so e.g., comparing only raw materials with each other. Input products with a higher carbon footprint face a higher risk.

The Input Product Carbon Transition Risk Indicator is therefore similar to the Product Carbon Transition Risk Indicator, but it focuses on the input products and not the product of the company. Input products are, for example, resources, packaging materials, energy and enabling services (such as tractor use on farm) to produce the product.

After identifying each carbon footprint for one input product, the input products are ranked according to their footprint. Input products in the highest percentile (≥70%) are classified as high transition risk input products. Input products in the medium percentile (between ≥30% and <70%) are classified as medium transition risk input products. Products in the lowest percentile (<30%) are classified as low transition risk input products.

After assessing the input products' transition risk based on the carbon footprint of each product, they are aggregated at the company-level. We derive what percentage of the input products are high, medium and low transition risk.

The transition risk indicator consists of 2 broad steps:

Score input products: Identifying the input products for each product, and calculating the relative carbon footprint per input product.

Score companies: Aggregating on the company-level.

The sample data set includes inputs and co2 footprints for each product from Ecoinvent and sectors from Europages. NOTE: the following columns are a completely random selection and do not reflect the true information:

co2 footprints (not allowed to share licensed data right now)

sectors (as the matching with ecoinvent is not done yet, we do not have one sector per product yet)

Usage

ictr(companies, inputs, low_threshold = 0.3, high_threshold = 0.7)

Arguments

companies

A data.frame like ictr_companies.

inputs

A data.frame like ictr_inputs

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

Value

A dataframe with columns id, transition_risk, and scores.

See also

Examples

ictr(ictr_companies, ictr_inputs)
#> # A tibble: 15 × 6
#>    id           transition_risk score_all score_unit score_sector score_unit_sec
#>    <chr>        <chr>               <dbl>      <dbl>        <dbl>          <dbl>
#>  1 fleischerei… high                0.286      0.333        0.333         0.476 
#>  2 fleischerei… medium              0.381      0.476        0.429         0.429 
#>  3 fleischerei… low                 0.333      0.190        0.238         0.0952
#>  4 pecheries-b… high                0.333      0.333        0.556         0.556 
#>  5 pecheries-b… medium              0.556      0.556        0.333         0.222 
#>  6 pecheries-b… low                 0.111      0.111        0.111         0.222 
#>  7 hoche-butte… high                0          0            0             0     
#>  8 hoche-butte… medium              0.667      0.667        0.667         0.667 
#>  9 hoche-butte… low                 0.333      0.333        0.333         0.333 
#> 10 vicquelin-e… high                0.111      0.222        0.111         0.222 
#> 11 vicquelin-e… medium              0.444      0.444        0.556         0.667 
#> 12 vicquelin-e… low                 0.444      0.333        0.333         0.111 
#> 13 bst-procont… high                0.394      0.364        0.455         0.515 
#> 14 bst-procont… medium              0.333      0.424        0.333         0.333 
#> 15 bst-procont… low                 0.273      0.212        0.212         0.152