Aggregate rank and score for each input product

Input products will be ranked according to their footprint. Input products in the highest percentile (≥70%) will be classified as high transition risk input products. Input products in the medium percentile (between ≥30% and <70%) will be classified as medium transition risk input products. Products in the lowest percentile (<30%) will be classified as low transition risk input products.

Therefore, each input product will receive either a score of 'high', 'medium', or 'low' based on the above mentioned classification criteria.

Usage

ictr_at_product_level(co2, low_threshold = 0.3, high_threshold = 0.7)

Arguments

co2

A data.frame like ictr_inputs.

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

Value

A data.frame.

See also

Author

Kalash Singhal.

Examples

ictr_at_product_level(ictr_inputs)
#> # A tibble: 75 × 12
#>    input_co2 input_sector        unit  activity_uuid_produc…¹ perc_all perc_unit
#>        <dbl> <chr>               <chr> <chr>                     <dbl>     <dbl>
#>  1      73.6 transport           metr… 0a242b09-772a-5edf-8e…    0.78      0.667
#>  2      96   infrastructure & m… kg    be06d25c-73dc-55fb-96…    0.947     0.956
#>  3      39.1 electricity         metr… 977d997e-c257-5033-ba…    0.387     0.444
#>  4      80.1 infrastructure & m… kg    5fd7fca0-0700-51fd-87…    0.82      0.878
#>  5      10.7 agriculture & anim… m     ebb8475e-ff57-5e4e-93…    0.187     0.667
#>  6      25.7 infrastructure & m… kg    2f7b77a7-1556-5c1b-b0…    0.293     0.267
#>  7      16.9 infrastructure & m… km    edd290d2-a5b0-56ec-aa…    0.247     0.333
#>  8      55.8 fuels               kg    d2c0844e-6e5b-52a3-80…    0.58      0.578
#>  9      29.1 infrastructure & m… kg    abdb78ee-ddb9-5552-9c…    0.36      0.356
#> 10      92.7 infrastructure & m… kg    16e1dc41-2793-521c-b4…    0.887     0.933
#> # ℹ 65 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 6 more variables: perc_sec <dbl>, perc_unit_sec <dbl>, score_all <chr>,
#> #   score_unit <chr>, score_sector <chr>, score_unit_sec <chr>