Aggregate the input products' scores for each company

Calculates on a company-level the percentage of input products that are in low / medium / high transition risk due to the input products' relative carbon footprint.

Activities are mapped with the companies' input products using common columns present in dataframes ecoinvent_scores and companies.

Usage

ictr_score_companies(ecoinvent_scores, companies)

Arguments

ecoinvent_scores

A data.frame. The output of ictr_score_inputs.

companies

A data.frame like ictr_companies.

Value

A data.frame with three rows for each company and these columns:

company_id

score

share_all

share_unit

share_sector

share_unit_sec where share_all, share_unit, share_sector, and share_unit_sec holds the aggregated scores in percentage.

See also

Author

Kalash Singhal.

Examples

library(dplyr, warn.conflicts = FALSE)

one_company <- ictr_companies |> filter(company_id %in% first(company_id))
one_company
#> # A tibble: 7 × 4
#>   activity_product_uuid                             company_id ei_activity unit 
#>   <chr>                                             <chr>      <chr>       <chr>
#> 1 0a242b09-772a-5edf-8e82-9cb4ba52a258_ae39ee61-d4… fleischer… transport,… metr…
#> 2 be06d25c-73dc-55fb-965b-0f300453e380_98b48ff2-22… fleischer… treatment … kg   
#> 3 977d997e-c257-5033-ba39-d0edeeef4ba2_0ace02fa-ec… fleischer… market gro… metr…
#> 4 5fd7fca0-0700-51fd-877a-72b5b6ff8868_9219c94a-7a… fleischer… treatment … kg   
#> 5 ebb8475e-ff57-5e4e-937b-b5788186a5ca_ccee034c-8b… fleischer… market for… m    
#> 6 2f7b77a7-1556-5c1b-b0aa-c4534ddc8885_38d493e9-6f… fleischer… market for… kg   
#> 7 edd290d2-a5b0-56ec-aaf3-ad71ce5f9403_095528f2-b5… fleischer… transport,… km   

ictr_inputs |>
  ictr_score_inputs() |>
  ictr_score_companies(one_company)
#> # A tibble: 3 × 6
#>   company_id              score share_all share_unit share_sector share_unit_sec
#>   <chr>                   <chr>     <dbl>      <dbl>        <dbl>          <dbl>
#> 1 fleischerei-stiefsohn_… high      0.286      0.333        0.333         0.476 
#> 2 fleischerei-stiefsohn_… medi…     0.381      0.476        0.429         0.429 
#> 3 fleischerei-stiefsohn_… low       0.333      0.190        0.238         0.0952