Calculate the ISTR indicator

The Input Product Sector Risk Indicator assesses the transition risk of the input products based on the sector's emissions targets the input product belongs to. Those sector emission reduction targets vary across scenarios (e.g., net zero scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040, 2050 to achieve the targets). It, therefore, is similar to the Product Sector Risk Indicator but focuses on the input products that the company needs to produce its products.The input products are, for example, resources, packaging materials, energy and enabling services (such as tractor use on farm) to produce the product.

If an input product belongs to a sector in which reduction targets are over 70%, it belongs to a high-risk sector and is classified as 'high sector transition risk'. If the input product belongs to a sector in which reduction targets are between 30-70%, the product is classified into the category 'medium sector transition risk' and if it is below 30% it belongs to the category of 'low sector transition risk'. If it does not belong to a sector that has decarbonization targets, it is classified as 'no sector transition risk'.

After assessing the input products for each product, they are aggregated at company-level to derive what percentage of the input products required by the company to produce its products have high, medium and low sector transition risk. We, therefore, derive the company-level information.

Please note that carbon emissions or emissions always mean CO2e.

Usage

istr(companies, scenario, mapper)

Arguments

companies

A data.frame. The output of istr_mapping().

scenario

A dataframe with scenario data.

mapper

A dataframe mapping companies to scenario.

Value

A dataframe with columns id, transition_risk, and scores.

See also

Examples

istr(istr_companies, istr_weo_2022, istr_ep_weo)
#> Warning: Returning more (or less) than 1 row per `summarise()` group was deprecated in
#> dplyr 1.1.0.
#> ℹ Please use `reframe()` instead.
#> ℹ When switching from `summarise()` to `reframe()`, remember that `reframe()`
#>   always returns an ungrouped data frame and adjust accordingly.
#> ℹ The deprecated feature was likely used in the tiltIndicator package.
#>   Please report the issue at
#>   <https://github.com/2DegreesInvesting/tiltIndicator/issues>.
#> # A tibble: 16 × 6
#>    id               transition_risk score_aggregated company_name scenario  year
#>    <chr>            <chr>                      <dbl> <chr>        <chr>    <dbl>
#>  1 fleischerei-sti… high                       400   fleischerei… Net Zer…  2040
#>  2 fleischerei-sti… high                      2467.  fleischerei… Net Zer…  2050
#>  3 fleischerei-sti… low                       2467.  fleischerei… Stated …  2020
#>  4 fleischerei-sti… medium                    2467.  fleischerei… Net Zer…  2030
#>  5 fleischerei-sti… medium                    2067.  fleischerei… Net Zer…  2040
#>  6 fleischerei-sti… no_sector                   28.6 fleischerei… NA          NA
#>  7 hoche-butter-gm… high                       467.  hoche butte… Net Zer…  2040
#>  8 hoche-butter-gm… high                      3867.  hoche butte… Net Zer…  2050
#>  9 hoche-butter-gm… low                       3867.  hoche butte… Stated …  2020
#> 10 hoche-butter-gm… medium                    3867.  hoche butte… Net Zer…  2030
#> 11 hoche-butter-gm… medium                    3400   hoche butte… Net Zer…  2040
#> 12 pecheries-basqu… high                      5911.  pecheries b… Net Zer…  2050
#> 13 pecheries-basqu… low                       5911.  pecheries b… Stated …  2020
#> 14 pecheries-basqu… medium                    5911.  pecheries b… Net Zer…  2030
#> 15 pecheries-basqu… medium                    5911.  pecheries b… Net Zer…  2040
#> 16 pecheries-basqu… no_sector                   22.2 pecheries b… NA          NA