Aggregate the products' scores for each company

Calculates on a company-level the percentage of products that are in low / medium / high transition risk.

Usage

istr_aggregate_scores(with_transition_risk, companies)

Arguments

with_transition_risk

A data.frame. The output of istr_add_transition_risk().

companies

A data.frame like istr_companies.

Value

A data.frame with the columns:

company_name

transition_risk

scenario

year

score_aggregated, which holds the aggregated scores in percentage.

See also

Author

Lyanne Ho.

Examples

istr_companies |>
  istr_mapping(istr_ep_weo) |>
  istr_add_reductions(istr_weo_2022) |>
  istr_add_transition_risk() |>
  istr_aggregate_scores(istr_companies)
#> Warning: Returning more (or less) than 1 row per `summarise()` group was deprecated in
#> dplyr 1.1.0.
#> ℹ Please use `reframe()` instead.
#> ℹ When switching from `summarise()` to `reframe()`, remember that `reframe()`
#>   always returns an ungrouped data frame and adjust accordingly.
#> ℹ The deprecated feature was likely used in the tiltIndicator package.
#>   Please report the issue at
#>   <https://github.com/2DegreesInvesting/tiltIndicator/issues>.
#> # A tibble: 16 × 5
#> # Groups:   company_name, transition_risk, scenario, year [16]
#>    company_name          transition_risk scenario          year score_aggregated
#>    <chr>                 <chr>           <chr>            <dbl>            <dbl>
#>  1 fleischerei stiefsohn high            Net Zero Emissi…  2040            400  
#>  2 fleischerei stiefsohn high            Net Zero Emissi…  2050           2467. 
#>  3 fleischerei stiefsohn low             Stated Policies…  2020           2467. 
#>  4 fleischerei stiefsohn medium          Net Zero Emissi…  2030           2467. 
#>  5 fleischerei stiefsohn medium          Net Zero Emissi…  2040           2067. 
#>  6 fleischerei stiefsohn no_sector       NA                  NA             28.6
#>  7 hoche butter gmbh     high            Net Zero Emissi…  2040            467. 
#>  8 hoche butter gmbh     high            Net Zero Emissi…  2050           3867. 
#>  9 hoche butter gmbh     low             Stated Policies…  2020           3867. 
#> 10 hoche butter gmbh     medium          Net Zero Emissi…  2030           3867. 
#> 11 hoche butter gmbh     medium          Net Zero Emissi…  2040           3400  
#> 12 pecheries basques     high            Net Zero Emissi…  2050           5911. 
#> 13 pecheries basques     low             Stated Policies…  2020           5911. 
#> 14 pecheries basques     medium          Net Zero Emissi…  2030           5911. 
#> 15 pecheries basques     medium          Net Zero Emissi…  2040           5911. 
#> 16 pecheries basques     no_sector       NA                  NA             22.2