Calculate the PCTR indicator

The Product Carbon Transition Risk measures the transition risk on the product-level of a company. The indicator is expressed in the percentage of all products that are at high risk, medium risk or low risk due to the products' relative carbon footprint. The assessment is first performed on a product-level and can then be aggregated to the company-level.

The Product Carbon Transition Risk measures the relative carbon footprint per product. As a default option each product is compared to the carbon footprint of every other product. Alternatively, users can also choose to compare those products of the same type, so e.g., comparing only raw materials with each other. Products with a higher carbon-footprint face a higher risk.

After identifying each carbon footprint for one product, the products are ranked according to their carbon footprint. Products in the highest percentile (≥70%) are classified as high transition risk products. Products in the medium percentile (between ≥30% and <70%) are classified as medium transition risk products. Products in the lowest percentile (<30%) are classified as low transition risk products.

After categorization, we aggregate all products from the same category and set them in relation to all products that the company produces. We derive the Product Carbon Transition Risks.

Please note that carbon footprints, and emissions are used equivalently. Carbon footprint refers only on the emissions which occur at the production stage of the product and not the emissions from the inputs. The unit is CO2e in kg.

This indicator provides the share of products with "low", "medium", or "high" relative production emissions per company.

The output of this indicator contains the following:

a column with production emissions

a column indicating the percentile relative to (i) all products with same unit (ii) products in same sector (iii) products in same segment

a column indicating whether the product has "low", "medium" or "high" relative production emissions.

Usage

pctr(companies, co2, low_threshold = 0.3, high_threshold = 0.7)

pctr_at_product_level(
  companies,
  co2,
  low_threshold = 0.3,
  high_threshold = 0.7
)

pctr_at_company_level(data)

Arguments

companies

A dataframe like pctr_companies.

co2

A dataframe like pctr_ecoinvent_co2.

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

data

A dataframe. The ouput of ictr_at_product_level().

Value

A dataframe with columns id, transition_risk, and scores.

Examples

companies <- pctr_companies
co2 <- pctr_ecoinvent_co2

# Product level
companies |>
  pctr_at_product_level(co2)
#> # A tibble: 25 × 14
#>    activity_product_uuid     company_id ei_activity.x unit.x co2_footprint sec  
#>    <chr>                     <chr>      <chr>         <chr>          <dbl> <chr>
#>  1 0a242b09-772a-5edf-8e82-… powerstar… transport, f… metri…          4.33 Tran…
#>  2 be06d25c-73dc-55fb-965b-… powerstar… treatment of… kg              3.94 Wast…
#>  3 977d997e-c257-5033-ba39-… powerstar… market group… metri…          6.26 Tran…
#>  4 5fd7fca0-0700-51fd-877a-… powerstar… treatment of… kg              6.99 Wast…
#>  5 ebb8475e-ff57-5e4e-937b-… ebtor_000… market for v… m               7.83 Infr…
#>  6 2f7b77a7-1556-5c1b-b0aa-… ebtor_000… market for s… kg              5.97 Chem…
#>  7 edd290d2-a5b0-56ec-aaf3-… ebtor_000… transport, p… km              9.00 Tran…
#>  8 d2c0844e-6e5b-52a3-802c-… ebtor_000… market for m… kg              5.10 Agri…
#>  9 abdb78ee-ddb9-5552-9c4c-… ebtor_000… citric acid … kg              7.50 Chem…
#> 10 16e1dc41-2793-521c-b4c9-… mbdgmbh_0… sodium ethyl… kg              6.60 Chem…
#> # ℹ 15 more rows
#> # ℹ 8 more variables: unit.y <chr>, ei_activity.y <chr>, perc_all <dbl>,
#> #   perc_unit <dbl>, perc_unit_sec <dbl>, score_all <chr>, score_unit <chr>,
#> #   score_unit_sec <chr>

# Company level
companies |>
  pctr_at_product_level(co2) |>
  pctr_at_company_level()
#> # A tibble: 9 × 5
#>   company_id                           score share_all share_unit share_unit_sec
#>   <chr>                                <chr>     <dbl>      <dbl>          <dbl>
#> 1 powerstartmanufacturinggmbh_0000000… high       0.25      0.5           0.5   
#> 2 powerstartmanufacturinggmbh_0000000… medi…      0.25      0.25          0.25  
#> 3 powerstartmanufacturinggmbh_0000000… low        0.5       0.25          0.25  
#> 4 ebtor_00000005324899621837001        high       0.6       0.6           0.8   
#> 5 ebtor_00000005324899621837001        medi…      0.2       0.4           0.2   
#> 6 ebtor_00000005324899621837001        low        0.2       0             0     
#> 7 mbdgmbh_00000004773428001            high       0.25      0.438         0.688 
#> 8 mbdgmbh_00000004773428001            medi…      0.5       0.375         0.25  
#> 9 mbdgmbh_00000004773428001            low        0.25      0.188         0.0625

# Similar
pctr(companies, co2)
#> # A tibble: 9 × 5
#>   id                         transition_risk score_all score_unit score_unit_sec
#>   <chr>                      <chr>               <dbl>      <dbl>          <dbl>
#> 1 powerstartmanufacturinggm… high                 0.25      0.5           0.5   
#> 2 powerstartmanufacturinggm… medium               0.25      0.25          0.25  
#> 3 powerstartmanufacturinggm… low                  0.5       0.25          0.25  
#> 4 ebtor_0000000532489962183… high                 0.6       0.6           0.8   
#> 5 ebtor_0000000532489962183… medium               0.2       0.4           0.2   
#> 6 ebtor_0000000532489962183… low                  0.2       0             0     
#> 7 mbdgmbh_00000004773428001  high                 0.25      0.438         0.688 
#> 8 mbdgmbh_00000004773428001  medium               0.5       0.375         0.25  
#> 9 mbdgmbh_00000004773428001  low                  0.25      0.188         0.0625