Aggregate rank and score for each activity

Activities will be ranked according to their carbon footprint. Activities in the highest percentile (≥70%) will be classified as high transition risk products. Activities in the medium percentile (between ≥30% and <70%) will be classified as medium transition risk products. Activities in the lowest percentile (<30%) will be classified as low transition risk products.

Therefore, each activity will receive either a score of 'high', 'medium', or 'low' based on the above mentioned classification criteria.

Usage

pctr_score_activities(ecoinvent_co2, low_threshold = 0.3, high_threshold = 0.7)

Arguments

ecoinvent_co2

A data.frame like pctr_ecoinvent_co2.

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

Value

A data.frame with columns:

All the columns from the ecoinvent_co2 dataset.

New columns:

perc_all

perc_unit

perc_unit_sec

score_all

score_unit

score_unit_sec

See also

Author

Tilman Trompke.

Examples

pctr_ecoinvent_co2 |>
  pctr_score_activities()
#> # A tibble: 25 × 11
#>    co2_footprint sec            unit  activity_product_uuid ei_activity perc_all
#>            <dbl> <chr>          <chr> <chr>                 <chr>          <dbl>
#>  1          4.33 Transport      metr… 0a242b09-772a-5edf-8… transport,…     0.08
#>  2          3.94 Waste Treatme… kg    be06d25c-73dc-55fb-9… treatment …     0.04
#>  3          6.26 Transport      metr… 977d997e-c257-5033-b… market gro…     0.68
#>  4          6.99 Waste Treatme… kg    5fd7fca0-0700-51fd-8… treatment …     0.88
#>  5          7.83 Infrastructur… m     ebb8475e-ff57-5e4e-9… market for…     0.96
#>  6          5.97 Chemicals      kg    2f7b77a7-1556-5c1b-b… market for…     0.6 
#>  7          9.00 Transport      km    edd290d2-a5b0-56ec-a… transport,…     1   
#>  8          5.10 Agriculture &… kg    d2c0844e-6e5b-52a3-8… market for…     0.28
#>  9          7.50 Chemicals      kg    abdb78ee-ddb9-5552-9… citric aci…     0.92
#> 10          6.60 Chemicals      kg    16e1dc41-2793-521c-b… sodium eth…     0.72
#> # ℹ 15 more rows
#> # ℹ 5 more variables: perc_unit <dbl>, perc_unit_sec <dbl>, score_all <chr>,
#> #   score_unit <chr>, score_unit_sec <chr>