Calculate the PSTR indicator

The Product Sector Transition Risk Indicator measures the transition risk of products based on the sector's emissions targets the product belongs to. Those sector emission reduction targets vary across scenarios (e.g., net zero scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are aggregated and set in relation to all products of the company. We, therefore, derive company-level information.

Usage

pstr(companies, scenarios)

pstr_at_product_level(companies, scenarios)

pstr_at_company_level(data, companies)

Arguments

companies

A dataframe like pstr_companies.

scenarios

A dataframe like pstr_scenarios.

data

A dataframe. The output at product level.

Value

At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.

Examples

companies <- pstr_companies
scenarios <- pstr_scenarios

# Product level
companies |>
  pstr_at_product_level(scenarios)
#> # A tibble: 126 × 19
#>    companies_id       grouped_by risk_category company_name products isic_4digit
#>    <chr>              <chr>      <chr>         <chr>        <chr>          <dbl>
#>  1 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  2 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  3 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  4 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  5 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  6 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  7 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  8 cta-commodity-tra… 1.5c requ… low           cta - commo… NA                NA
#>  9 cta-commodity-tra… NA_NA      no_sector     cta - commo… NA                NA
#> 10 cbrit_00000005227… 1.5c requ… low           cbr-it       NA                NA
#> # ℹ 116 more rows
#> # ℹ 13 more variables: tilt_sector <chr>, tilt_subsector <chr>, type <chr>,
#> #   sector <chr>, subsector <chr>, scenario <chr>, region <chr>, year <dbl>,
#> #   value <dbl>, reductions <dbl>, publication <chr>, product <chr>, flow <chr>

# Company level
companies |>
  pstr_at_product_level(scenarios) |>
  pstr_at_company_level(companies)
#> # A tibble: 45 × 6
#>    companies_id                    grouped_by risk_category value scenario  year
#>    <chr>                           <chr>      <chr>         <dbl> <chr>    <dbl>
#>  1 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2020
#>  2 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2030
#>  3 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2040
#>  4 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2050
#>  5 barham-metall-gmbh_00000005371… NA_NA      no_sector        50 NA          NA
#>  6 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2020
#>  7 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2030
#>  8 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2040
#>  9 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2050
#> 10 breuninger-leder-gmbh_00000005… NA_NA      no_sector        50 NA          NA
#> # ℹ 35 more rows

# Same
pstr(companies, scenarios)
#> # A tibble: 45 × 6
#>    companies_id                    grouped_by risk_category value scenario  year
#>    <chr>                           <chr>      <chr>         <dbl> <chr>    <dbl>
#>  1 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2020
#>  2 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2030
#>  3 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2040
#>  4 barham-metall-gmbh_00000005371… 1.5c requ… low             100 1.5c re…  2050
#>  5 barham-metall-gmbh_00000005371… NA_NA      no_sector        50 NA          NA
#>  6 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2020
#>  7 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2030
#>  8 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2040
#>  9 breuninger-leder-gmbh_00000005… 1.5c requ… low             100 1.5c re…  2050
#> 10 breuninger-leder-gmbh_00000005… NA_NA      no_sector        50 NA          NA
#> # ℹ 35 more rows