Given a named list of scenarios returns a cleaner scenarios dataframe

Given a named list of scenarios returns a cleaner scenarios dataframe

Usage

pstr_prepare_scenario(scenarios)

Arguments

scenarios

A named list of identically structured scenarios.

Value

A single, cleaner dataframe with an additional column to identify which rows come from which scenario.

See also

Examples

library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)

raw_weo <- read_csv(extdata_path("pstr_raw_weo_2022.csv"))
#> Rows: 372 Columns: 8
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (5): publication, scenario, region, weo_product, weo_flow
#> dbl (3): year, value, co2_reductions
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
raw_ipr <- read_csv(extdata_path("pstr_raw_ipr_2022.csv"))
#> Rows: 112 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (4): scenario, region, ipr_sector, ipr_subsector
#> dbl (3): year, value, co2_reductions
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
raw_scenarios <- list(weo = raw_weo, ipr = raw_ipr)

pstr_prepare_scenario(raw_scenarios)
#> # A tibble: 484 × 11
#>    publication       scenario region product flow   year  value reductions type 
#>    <chr>             <chr>    <chr>  <chr>   <chr> <dbl>  <dbl>      <dbl> <chr>
#>  1 world energy out… stated … world  total   tota…  2020 31904      0      weo  
#>  2 world energy out… stated … world  total   tota…  2030 33135.    -0.0386 weo  
#>  3 world energy out… stated … world  total   tota…  2040 30800.     0.0346 weo  
#>  4 world energy out… stated … world  total   tota…  2050 28946.     0.0927 weo  
#>  5 world energy out… stated … world  coal    tota…  2020 14335.     0      weo  
#>  6 world energy out… stated … world  coal    tota…  2030 13695      0.0447 weo  
#>  7 world energy out… stated … world  coal    tota…  2040 11553.     0.194  weo  
#>  8 world energy out… stated … world  coal    tota…  2050  9863.     0.312  weo  
#>  9 world energy out… stated … world  oil     tota…  2020 10194.     0      weo  
#> 10 world energy out… stated … world  oil     tota…  2030 11412.    -0.119  weo  
#> # ℹ 474 more rows
#> # ℹ 2 more variables: sector <chr>, subsector <chr>