Calculate the indicator "sector profile"

The indicator "sector profile" measures the transition risk of products based on the sector's emissions targets the product belongs to. Those sector emission reduction targets vary across scenarios (e.g., net zero scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are aggregated and set in relation to all products of the company. We, therefore, derive company-level information.

Usage

sector_profile(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

Arguments

companies, scenarios

A dataframe like the dataset with a matching name in tiltToyData (see Reference).

low_threshold

A numeric value to segment low and medium reduction targets.

high_threshold

A numeric value to segment medium and high reduction targets.

Value

A data frame with the column companies_id, and the nested columnsproduct and company holding the outputs at product and company level. Unnesting product yields a data frame with at least columns companies_id, grouped_by, risk_category. Unnesting company yields a data frame with at least columns companies_id, grouped_by, risk_category, value. Any column in the input datasets ending with *rowid is also passed as is to the output at product level. The exception is any column named exactly rowid-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.

See also

Examples

library(tiltIndicator)
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_sector_profile_companies())
scenarios <- read_csv(toy_sector_profile_any_scenarios())

both <- sector_profile(companies, scenarios)
both
#> # A tibble: 72 × 3
#>    companies_id                       product           company          
#>    <chr>                              <list>            <list>           
#>  1 antimonarchy_canine                <tibble [4 × 10]> <tibble [12 × 3]>
#>  2 celestial_lovebird                 <tibble [4 × 10]> <tibble [12 × 3]>
#>  3 nonphilosophical_llama             <tibble [8 × 10]> <tibble [12 × 3]>
#>  4 asteria_megalotomusquinquespinosus <tibble [4 × 10]> <tibble [12 × 3]>
#>  5 quasifaithful_amphiuma             <tibble [4 × 10]> <tibble [12 × 3]>
#>  6 spectacular_americanriverotter     <tibble [4 × 10]> <tibble [12 × 3]>
#>  7 contrite_silkworm                  <tibble [4 × 10]> <tibble [12 × 3]>
#>  8 harmless_owlbutterfly              <tibble [4 × 10]> <tibble [12 × 3]>
#>  9 fascist_maiasaura                  <tibble [4 × 10]> <tibble [12 × 3]>
#> 10 charismatic_islandwhistler         <tibble [4 × 10]> <tibble [12 × 3]>
#> # ℹ 62 more rows

both |> unnest_product()
#> # A tibble: 304 × 11
#>    companies_id           grouped_by     risk_category profile_ranking clustered
#>    <chr>                  <chr>          <chr>                   <dbl> <chr>    
#>  1 antimonarchy_canine    ipr_1.5C RPS_… medium                   0.18 tent     
#>  2 antimonarchy_canine    ipr_1.5C RPS_… high                     0.98 tent     
#>  3 antimonarchy_canine    weo_NZ 2050_2… high                     0.4  tent     
#>  4 antimonarchy_canine    weo_NZ 2050_2… high                     0.97 tent     
#>  5 celestial_lovebird     ipr_1.5C RPS_… medium                   0.18 table hi…
#>  6 celestial_lovebird     ipr_1.5C RPS_… high                     0.98 table hi…
#>  7 celestial_lovebird     weo_NZ 2050_2… high                     0.4  table hi…
#>  8 celestial_lovebird     weo_NZ 2050_2… high                     0.97 table hi…
#>  9 nonphilosophical_llama ipr_1.5C RPS_… low                      0.09 surface …
#> 10 nonphilosophical_llama ipr_1.5C RPS_… high                     0.95 surface …
#> # ℹ 294 more rows
#> # ℹ 6 more variables: activity_uuid_product_uuid <chr>, tilt_sector <chr>,
#> #   scenario <chr>, year <dbl>, type <chr>, tilt_subsector <chr>

both |> unnest_company()
#> # A tibble: 864 × 4
#>    companies_id        grouped_by        risk_category value
#>    <chr>               <chr>             <chr>         <dbl>
#>  1 antimonarchy_canine ipr_1.5C RPS_2030 high              0
#>  2 antimonarchy_canine ipr_1.5C RPS_2030 medium            1
#>  3 antimonarchy_canine ipr_1.5C RPS_2030 low               0
#>  4 antimonarchy_canine ipr_1.5C RPS_2050 high              1
#>  5 antimonarchy_canine ipr_1.5C RPS_2050 medium            0
#>  6 antimonarchy_canine ipr_1.5C RPS_2050 low               0
#>  7 antimonarchy_canine weo_NZ 2050_2030  high              1
#>  8 antimonarchy_canine weo_NZ 2050_2030  medium            0
#>  9 antimonarchy_canine weo_NZ 2050_2030  low               0
#> 10 antimonarchy_canine weo_NZ 2050_2050  high              1
#> # ℹ 854 more rows