Calculate input (or product) carbon transition risk

These functions calculate the input (or product) carbon transition risk. The process is the same. What varies is the co2 dataset.

Usage

xctr(companies, co2, low_threshold = 1/3, high_threshold = 2/3)

xctr_at_company_level(data)

xctr_at_product_level(
  companies,
  co2,
  low_threshold = 1/3,
  high_threshold = 2/3
)

Arguments

companies

A dataframe like companies.

co2

A dataframe like inputs.

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

data

A dataframe. The output at product level.

Value

At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.

Details

Depracated

The ictr*() and pctr*() functions are now deprecated. Use the xctr*() functions instead.

Input carbon transition risk (ICTR)

The Input Product Carbon Transition Risk Indicator assesses the transition risk of the input products due to their relative carbon footprint to other input products. As a default option, each input product is compared to the carbon footprint of every other input product. Input products with a higher carbon footprint face a higher risk.

The Input Product Carbon Transition Risk Indicator is therefore similar to the Product Carbon Transition Risk Indicator, but it focuses on the input products and not the product of the company. Input products are, for example, resources, packaging materials, energy and enabling services (such as tractor use on farm) to produce the product.

After identifying each carbon footprint for one input product, the input products are ranked according to their footprint. Input products in the highest percentile (≥70%) are classified as high transition risk input products. Input products in the medium percentile (between ≥30% and <70%) are classified as medium transition risk input products. Products in the lowest percentile (<30%) are classified as low transition risk input products.

After assessing the input products' transition risk based on the carbon footprint of each product, they are aggregated at the company-level. We derive what percentage of the input products are high, medium and low transition risk.

The transition risk indicator consists of 2 broad steps:

Score input products: Identifying the input products for each product, and calculating the relative carbon footprint per input product.

Score companies: Aggregating on the company-level.

The sample data set includes inputs and co2 footprints for each product from Ecoinvent and sectors from Europages. NOTE: the following columns are a completely random selection and do not reflect the true information:

co2 footprints (not allowed to share licensed data right now)

sectors (as the matching with ecoinvent is not done yet, we do not have one sector per product yet)

Product carbon transition risk (PCTR)

The Product Carbon Transition Risk measures the transition risk on the product-level of a company. The indicator is expressed in the percentage of all products that are at high risk, medium risk or low risk due to the products' relative carbon footprint. The assessment is first performed on a product-level and can then be aggregated to the company-level.

The Product Carbon Transition Risk measures the relative carbon footprint per product. As a default option each product is compared to the carbon footprint of every other product. Products with a higher carbon-footprint face a higher risk.

After identifying each carbon footprint for one product, the products are ranked according to their carbon footprint. Products in the highest percentile (≥70%) are classified as high transition risk products. Products in the medium percentile (between ≥30% and <70%) are classified as medium transition risk products. Products in the lowest percentile (<30%) are classified as low transition risk products.

After categorization, we aggregate all products from the same category and set them in relation to all products that the company produces. We derive the Product Carbon Transition Risks.

Please note that carbon footprints, and emissions are used equivalently. Carbon footprint refers only on the emissions which occur at the production stage of the product and not the emissions from the inputs. The unit is CO2e in kg.

This indicator provides the share of products with "low", "medium", or "high" relative production emissions per company.

The output of this indicator contains the following:

a column with production emissions

a column indicating the percentile relative to (i) all products with same unit (ii) products in same sector (iii) products in same segment

a column indicating whether the product has "low", "medium" or "high" relative production emissions.

Examples

companies <- companies

# ICTR
inputs <- inputs

companies |>
  xctr_at_product_level(co2 = inputs)
#> # A tibble: 319 × 7
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… all        high          stove     0a242b09-772a-5edf-8e…
#>  2 fleischerei-stiefs… isic_sec   high          stove     0a242b09-772a-5edf-8e…
#>  3 fleischerei-stiefs… tilt_sec   high          stove     0a242b09-772a-5edf-8e…
#>  4 fleischerei-stiefs… unit       high          stove     0a242b09-772a-5edf-8e…
#>  5 fleischerei-stiefs… unit_isic… high          stove     0a242b09-772a-5edf-8e…
#>  6 fleischerei-stiefs… unit_tilt… high          stove     0a242b09-772a-5edf-8e…
#>  7 fleischerei-stiefs… all        high          stove     0a242b09-772a-5edf-8e…
#>  8 fleischerei-stiefs… isic_sec   high          stove     0a242b09-772a-5edf-8e…
#>  9 fleischerei-stiefs… tilt_sec   high          stove     0a242b09-772a-5edf-8e…
#> 10 fleischerei-stiefs… unit       high          stove     0a242b09-772a-5edf-8e…
#> # ℹ 309 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 2 more variables: input_activity_uuid_product_uuid <chr>,
#> #   input_co2_footprint <dbl>

companies |>
  xctr_at_product_level(co2 = inputs) |>
  xctr_at_company_level()
#> # A tibble: 126 × 4
#>    companies_id                           grouped_by risk_category value
#>    <chr>                                  <chr>      <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 all        high          0.143
#>  2 bst-procontrol-gmbh_00000005104947-001 all        medium        0.429
#>  3 bst-procontrol-gmbh_00000005104947-001 all        low           0.429
#>  4 bst-procontrol-gmbh_00000005104947-001 isic_sec   high          0.429
#>  5 bst-procontrol-gmbh_00000005104947-001 isic_sec   medium        0.286
#>  6 bst-procontrol-gmbh_00000005104947-001 isic_sec   low           0.286
#>  7 bst-procontrol-gmbh_00000005104947-001 tilt_sec   high          0.429
#>  8 bst-procontrol-gmbh_00000005104947-001 tilt_sec   medium        0.286
#>  9 bst-procontrol-gmbh_00000005104947-001 tilt_sec   low           0.286
#> 10 bst-procontrol-gmbh_00000005104947-001 unit       high          0.286
#> # ℹ 116 more rows

# Same
companies |> xctr(co2 = inputs)
#> # A tibble: 126 × 4
#>    companies_id                           grouped_by risk_category value
#>    <chr>                                  <chr>      <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 all        high          0.143
#>  2 bst-procontrol-gmbh_00000005104947-001 all        medium        0.429
#>  3 bst-procontrol-gmbh_00000005104947-001 all        low           0.429
#>  4 bst-procontrol-gmbh_00000005104947-001 isic_sec   high          0.429
#>  5 bst-procontrol-gmbh_00000005104947-001 isic_sec   medium        0.286
#>  6 bst-procontrol-gmbh_00000005104947-001 isic_sec   low           0.286
#>  7 bst-procontrol-gmbh_00000005104947-001 tilt_sec   high          0.429
#>  8 bst-procontrol-gmbh_00000005104947-001 tilt_sec   medium        0.286
#>  9 bst-procontrol-gmbh_00000005104947-001 tilt_sec   low           0.286
#> 10 bst-procontrol-gmbh_00000005104947-001 unit       high          0.286
#> # ℹ 116 more rows

# PCTR
products <- products

companies |>
  xctr_at_product_level(co2 = products)
#> # A tibble: 79 × 6
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… all        high          stove     0a242b09-772a-5edf-8e…
#>  2 fleischerei-stiefs… isic_sec   high          stove     0a242b09-772a-5edf-8e…
#>  3 fleischerei-stiefs… tilt_sec   high          stove     0a242b09-772a-5edf-8e…
#>  4 fleischerei-stiefs… unit       high          stove     0a242b09-772a-5edf-8e…
#>  5 fleischerei-stiefs… unit_isic… high          stove     0a242b09-772a-5edf-8e…
#>  6 fleischerei-stiefs… unit_tilt… high          stove     0a242b09-772a-5edf-8e…
#>  7 fleischerei-stiefs… all        high          oven      be06d25c-73dc-55fb-96…
#>  8 fleischerei-stiefs… isic_sec   medium        oven      be06d25c-73dc-55fb-96…
#>  9 fleischerei-stiefs… tilt_sec   high          oven      be06d25c-73dc-55fb-96…
#> 10 fleischerei-stiefs… unit       medium        oven      be06d25c-73dc-55fb-96…
#> # ℹ 69 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 1 more variable: co2_footprint <dbl>

companies |>
  xctr_at_product_level(co2 = products) |>
  xctr_at_company_level()
#> # A tibble: 126 × 4
#>    companies_id                           grouped_by risk_category value
#>    <chr>                                  <chr>      <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 all        high            0.5
#>  2 bst-procontrol-gmbh_00000005104947-001 all        medium          0.5
#>  3 bst-procontrol-gmbh_00000005104947-001 all        low             0  
#>  4 bst-procontrol-gmbh_00000005104947-001 isic_sec   high            0.5
#>  5 bst-procontrol-gmbh_00000005104947-001 isic_sec   medium          0.5
#>  6 bst-procontrol-gmbh_00000005104947-001 isic_sec   low             0  
#>  7 bst-procontrol-gmbh_00000005104947-001 tilt_sec   high            0.5
#>  8 bst-procontrol-gmbh_00000005104947-001 tilt_sec   medium          0.5
#>  9 bst-procontrol-gmbh_00000005104947-001 tilt_sec   low             0  
#> 10 bst-procontrol-gmbh_00000005104947-001 unit       high            1  
#> # ℹ 116 more rows

# Same
companies |> xctr(co2 = products)
#> # A tibble: 126 × 4
#>    companies_id                           grouped_by risk_category value
#>    <chr>                                  <chr>      <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 all        high            0.5
#>  2 bst-procontrol-gmbh_00000005104947-001 all        medium          0.5
#>  3 bst-procontrol-gmbh_00000005104947-001 all        low             0  
#>  4 bst-procontrol-gmbh_00000005104947-001 isic_sec   high            0.5
#>  5 bst-procontrol-gmbh_00000005104947-001 isic_sec   medium          0.5
#>  6 bst-procontrol-gmbh_00000005104947-001 isic_sec   low             0  
#>  7 bst-procontrol-gmbh_00000005104947-001 tilt_sec   high            0.5
#>  8 bst-procontrol-gmbh_00000005104947-001 tilt_sec   medium          0.5
#>  9 bst-procontrol-gmbh_00000005104947-001 tilt_sec   low             0  
#> 10 bst-procontrol-gmbh_00000005104947-001 unit       high            1  
#> # ℹ 116 more rows