Calculate the indicator "emissions profile"

The "emissions profile" measures the transition risk on the product-level of a company. The indicator is expressed in the percentage of all products that are at high risk, medium risk or low risk due to the products' relative carbon footprint. The assessment is first performed on a product-level and can then be aggregated to the company-level.

The "emissions profile" measures the relative carbon footprint per product. As a default option each product is compared to the carbon footprint of every other product. Products with a higher carbon-footprint face a higher risk.

After identifying each carbon footprint for one product, the products are ranked according to their carbon footprint. The ranking method is explained in the Thresholds section.

After categorization, we aggregate all products from the same category and set them in relation to all products that the company produces. We derive the "emissions profile".

Please note that carbon footprints, and emissions are used equivalently. Carbon footprint refers only on the emissions which occur at the production stage of the product and not the emissions from the inputs. The unit is CO2e in kg.

This indicator provides the share of products with "low", "medium", or "high" relative production emissions per company.

The output of this indicator contains the following:

a column with production emissions

a column indicating the percentile relative to (i) all products with same unit (ii) products in same sector (iii) products in same segment

a column indicating whether the product has "low", "medium" or "high" relative production emissions.

Usage

emissions_profile(companies, co2, low_threshold = 1/3, high_threshold = 2/3)

Arguments

companies, co2

A dataframe like the dataset with a matching name in tiltToyData (see Reference).

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

Value

A data frame with the column companies_id, and the nested columnsproduct and company holding the outputs at product and company level. Unnesting product yields a data frame with at least columns companies_id, grouped_by, risk_category. Unnesting company yields a data frame with at least columns companies_id, grouped_by, risk_category, value. Any column in the input datasets ending with *rowid is also passed as is to the output at product level. The exception is any column named exactly rowid-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.

See also

Examples

library(tiltIndicator)
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products())

both <- emissions_profile(companies, products)
both
#> # A tibble: 8 × 3
#>   companies_id                             product           company          
#>   <chr>                                    <list>            <list>           
#> 1 fleischerei-stiefsohn_00000005219477-001 <tibble [12 × 6]> <tibble [18 × 3]>
#> 2 pecheries-basques_fra316541-00101        <tibble [6 × 6]>  <tibble [18 × 3]>
#> 3 hoche-butter-gmbh_deu422723-693847001    <tibble [6 × 6]>  <tibble [18 × 3]>
#> 4 vicquelin-espaces-verts_fra697272-00101  <tibble [6 × 6]>  <tibble [18 × 3]>
#> 5 bst-procontrol-gmbh_00000005104947-001   <tibble [6 × 6]>  <tibble [18 × 3]>
#> 6 leider-gmbh_00000005064318-001           <tibble [6 × 6]>  <tibble [18 × 3]>
#> 7 cheries-baqu_neu316541-00101             <tibble [6 × 6]>  <tibble [18 × 3]>
#> 8 ca-coity-trg-aua-gmbh_00000384-001       <tibble [1 × 6]>  <tibble [3 × 3]> 

both |> unnest_product()
#> # A tibble: 49 × 7
#>    companies_id               grouped_by risk_category profile_ranking clustered
#>    <chr>                      <chr>      <chr>                   <dbl> <chr>    
#>  1 fleischerei-stiefsohn_000… all        high                    1     stove    
#>  2 fleischerei-stiefsohn_000… isic_4dig… high                    1     stove    
#>  3 fleischerei-stiefsohn_000… tilt_sect… high                    1     stove    
#>  4 fleischerei-stiefsohn_000… unit       high                    1     stove    
#>  5 fleischerei-stiefsohn_000… unit_isic… high                    1     stove    
#>  6 fleischerei-stiefsohn_000… unit_tilt… high                    1     stove    
#>  7 fleischerei-stiefsohn_000… all        high                    0.8   oven     
#>  8 fleischerei-stiefsohn_000… isic_4dig… medium                  0.5   oven     
#>  9 fleischerei-stiefsohn_000… tilt_sect… medium                  0.667 oven     
#> 10 fleischerei-stiefsohn_000… unit       medium                  0.5   oven     
#> # ℹ 39 more rows
#> # ℹ 2 more variables: activity_uuid_product_uuid <chr>, co2_footprint <dbl>

both |> unnest_company()
#> # A tibble: 129 × 4
#>    companies_id                             grouped_by  risk_category value
#>    <chr>                                    <chr>       <chr>         <dbl>
#>  1 fleischerei-stiefsohn_00000005219477-001 all         high            1  
#>  2 fleischerei-stiefsohn_00000005219477-001 all         medium          0  
#>  3 fleischerei-stiefsohn_00000005219477-001 all         low             0  
#>  4 fleischerei-stiefsohn_00000005219477-001 isic_4digit high            0.5
#>  5 fleischerei-stiefsohn_00000005219477-001 isic_4digit medium          0.5
#>  6 fleischerei-stiefsohn_00000005219477-001 isic_4digit low             0  
#>  7 fleischerei-stiefsohn_00000005219477-001 tilt_sector high            0.5
#>  8 fleischerei-stiefsohn_00000005219477-001 tilt_sector medium          0.5
#>  9 fleischerei-stiefsohn_00000005219477-001 tilt_sector low             0  
#> 10 fleischerei-stiefsohn_00000005219477-001 unit        high            0.5
#> # ℹ 119 more rows