Calculate the indicator "emissions profile"

The "emissions profile" measures the absolute GHG emissions of a product in comparison to a chosen benchmark of products. The assessment is first performed on a product-level and can then be aggregated to the company-level. The profile is expressed as the share of a company's products that are in a "high", "medium", or "low" emission profile category, based on the comparison to the benchmark. A higher emission profile indicates a larger impact on climate change compared to the benchmark and can therefore also be interpreted as one climate risk component to assess products, companies, or loan portfolios.

The "emission profile" is calculated in these steps:

The relative GHG emissions per product are collected from a Life-Cycle-Analysis (LCA) database by matching the products from our company dataset to the products from the LCA dataset.

All products are ranked according to their GHG emissions.

The products are grouped by the following benchmarks:

all: All products.

isic_4digit: All products within the same ISIC 4 digit code (example: 0112 Growing of rice).

tilt_sector: All products within the same tilt sector (example: agriculture).

unit: All products with the same unit (example: kg).

unit_isic_4digit: All products with the same unit within the same ISIC 4 digit section (example: kg + 0112 Growing of rice).

unit_tilt_sector: All products with the same unit within the same tilt sector (example: kg + agriculture).

For each benchmark, products are assigned to the emission profile category "low", "medium" or "high", depending on the GHG emissions arising from their production process in comparison to all other products within the same benchmark. For the assignment of the three categories, thresholds are used. Please find more information about the thresholds in the Thresholds section.

For the company-level results, we aggregate all products from the same category and benchmark and set them in relation to all products that the company produces. The company-level results are expressed as the company's share of products per category "low", "medium", and "high" in comparison to each benchmark.

The output of this indicator contains the following:

A column indicating the benchmark to which a product is compared.

A column indicating whether the product has "low", "medium" or "high" relative GHG emissions.

A column indicating the share of the products per category and benchmark".

Usage

emissions_profile(companies, co2, low_threshold = 1/3, high_threshold = 2/3)

Arguments

companies, co2

A dataframe like the dataset with a matching name in tiltToyData (see Reference).

low_threshold

A numeric value to segment low and medium emission profile products.

high_threshold

A numeric value to segment medium and high emission profile products.

Value

A data frame with the column companies_id, and the nested columnsproduct and company holding the outputs at product and company level. Unnesting product yields a data frame with at least columns companies_id, grouped_by, risk_category. Unnesting company yields a data frame with at least columns companies_id, grouped_by, risk_category, value. Any column in the input datasets ending with *rowid is also passed as is to the output at product level. The exception is any column named exactly rowid-- which is a reserved name and throws an error. Note this feature makes no sense at company level because potentially multiple rows in the input datasets are summarized into a single row in the output at company level.

See also

Examples

library(tiltIndicator)
library(tiltToyData)
library(readr)
options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products_ecoinvent())

both <- emissions_profile(companies, products)
both
#> # A tibble: 72 × 3
#>    companies_id                       product           company          
#>    <chr>                              <list>            <list>           
#>  1 antimonarchy_canine                <tibble [6 × 6]>  <tibble [18 × 3]>
#>  2 celestial_lovebird                 <tibble [6 × 6]>  <tibble [18 × 3]>
#>  3 nonphilosophical_llama             <tibble [12 × 6]> <tibble [18 × 3]>
#>  4 asteria_megalotomusquinquespinosus <tibble [6 × 6]>  <tibble [18 × 3]>
#>  5 quasifaithful_amphiuma             <tibble [6 × 6]>  <tibble [18 × 3]>
#>  6 spectacular_americanriverotter     <tibble [6 × 6]>  <tibble [18 × 3]>
#>  7 contrite_silkworm                  <tibble [6 × 6]>  <tibble [18 × 3]>
#>  8 harmless_owlbutterfly              <tibble [6 × 6]>  <tibble [18 × 3]>
#>  9 fascist_maiasaura                  <tibble [6 × 6]>  <tibble [18 × 3]>
#> 10 charismatic_islandwhistler         <tibble [6 × 6]>  <tibble [18 × 3]>
#> # ℹ 62 more rows

both |> unnest_product()
#> # A tibble: 456 × 7
#>    companies_id        grouped_by       risk_category profile_ranking clustered 
#>    <chr>               <chr>            <chr>                   <dbl> <chr>     
#>  1 antimonarchy_canine all              high                        1 tent      
#>  2 antimonarchy_canine isic_4digit      high                        1 tent      
#>  3 antimonarchy_canine tilt_sector      high                        1 tent      
#>  4 antimonarchy_canine unit             high                        1 tent      
#>  5 antimonarchy_canine unit_isic_4digit high                        1 tent      
#>  6 antimonarchy_canine unit_tilt_sector high                        1 tent      
#>  7 celestial_lovebird  all              high                        1 table hir…
#>  8 celestial_lovebird  isic_4digit      high                        1 table hir…
#>  9 celestial_lovebird  tilt_sector      high                        1 table hir…
#> 10 celestial_lovebird  unit             high                        1 table hir…
#> # ℹ 446 more rows
#> # ℹ 2 more variables: activity_uuid_product_uuid <chr>, co2_footprint <dbl>

both |> unnest_company()
#> # A tibble: 1,296 × 4
#>    companies_id        grouped_by  risk_category value
#>    <chr>               <chr>       <chr>         <dbl>
#>  1 antimonarchy_canine all         high              1
#>  2 antimonarchy_canine all         medium            0
#>  3 antimonarchy_canine all         low               0
#>  4 antimonarchy_canine isic_4digit high              1
#>  5 antimonarchy_canine isic_4digit medium            0
#>  6 antimonarchy_canine isic_4digit low               0
#>  7 antimonarchy_canine tilt_sector high              1
#>  8 antimonarchy_canine tilt_sector medium            0
#>  9 antimonarchy_canine tilt_sector low               0
#> 10 antimonarchy_canine unit        high              1
#> # ℹ 1,286 more rows