Calculate the ICTR indicator

The Input Product Carbon Transition Risk Indicator assesses the transition risk of the input products due to their relative carbon footprint to other input products. As a default option, each input product is compared to the carbon footprint of every other input product. Input products with a higher carbon footprint face a higher risk.

The Input Product Carbon Transition Risk Indicator is therefore similar to the Product Carbon Transition Risk Indicator, but it focuses on the input products and not the product of the company. Input products are, for example, resources, packaging materials, energy and enabling services (such as tractor use on farm) to produce the product.

After identifying each carbon footprint for one input product, the input products are ranked according to their footprint. Input products in the highest percentile (≥70%) are classified as high transition risk input products. Input products in the medium percentile (between ≥30% and <70%) are classified as medium transition risk input products. Products in the lowest percentile (<30%) are classified as low transition risk input products.

After assessing the input products' transition risk based on the carbon footprint of each product, they are aggregated at the company-level. We derive what percentage of the input products are high, medium and low transition risk.

The transition risk indicator consists of 2 broad steps:

Score input products: Identifying the input products for each product, and calculating the relative carbon footprint per input product.

Score companies: Aggregating on the company-level.

The sample data set includes inputs and co2 footprints for each product from Ecoinvent and sectors from Europages. NOTE: the following columns are a completely random selection and do not reflect the true information:

co2 footprints (not allowed to share licensed data right now)

sectors (as the matching with ecoinvent is not done yet, we do not have one sector per product yet)

Usage

ictr(companies, co2, low_threshold = 0.3, high_threshold = 0.7)

ictr_at_product_level(
  companies,
  co2,
  low_threshold = 0.3,
  high_threshold = 0.7
)

ictr_at_company_level(data)

Arguments

companies

A dataframe like ictr_companies.

co2

A dataframe like ictr_inputs.

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

data

A dataframe. The output at product level.

Value

At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.

Examples

companies <- ictr_companies
co2 <- ictr_inputs

# Product level
companies |>
  ictr_at_product_level(co2)
#> # A tibble: 324 × 17
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… all        high          stove     0a242b09-772a-5edf-8e…
#>  2 fleischerei-stiefs… isic_sec   high          stove     0a242b09-772a-5edf-8e…
#>  3 fleischerei-stiefs… tilt_sec   high          stove     0a242b09-772a-5edf-8e…
#>  4 fleischerei-stiefs… unit       high          stove     0a242b09-772a-5edf-8e…
#>  5 fleischerei-stiefs… unit_isic… high          stove     0a242b09-772a-5edf-8e…
#>  6 fleischerei-stiefs… unit_tilt… high          stove     0a242b09-772a-5edf-8e…
#>  7 fleischerei-stiefs… all        high          stove     0a242b09-772a-5edf-8e…
#>  8 fleischerei-stiefs… isic_sec   high          stove     0a242b09-772a-5edf-8e…
#>  9 fleischerei-stiefs… tilt_sec   high          stove     0a242b09-772a-5edf-8e…
#> 10 fleischerei-stiefs… unit       high          stove     0a242b09-772a-5edf-8e…
#> # ℹ 314 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 12 more variables: input_activity_uuid_product_uuid <chr>, unit <chr>,
#> #   input_co2_footprint <dbl>, input_tilt_sector <chr>, input_unit <chr>,
#> #   input_isic_4digit <dbl>, perc_all <dbl>, perc_isic_sec <dbl>,
#> #   perc_tilt_sec <dbl>, perc_unit <dbl>, perc_unit_isic_sec <dbl>,
#> #   perc_unit_tilt_sec <dbl>

# Company level
companies |>
  ictr_at_product_level(co2) |>
  ictr_at_company_level()
#> # A tibble: 144 × 4
#>    companies_id                           grouped_by risk_category value
#>    <chr>                                  <chr>      <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 all        high          0.143
#>  2 bst-procontrol-gmbh_00000005104947-001 all        medium        0.429
#>  3 bst-procontrol-gmbh_00000005104947-001 all        low           0.429
#>  4 bst-procontrol-gmbh_00000005104947-001 isic_sec   high          0.429
#>  5 bst-procontrol-gmbh_00000005104947-001 isic_sec   medium        0.286
#>  6 bst-procontrol-gmbh_00000005104947-001 isic_sec   low           0.286
#>  7 bst-procontrol-gmbh_00000005104947-001 tilt_sec   high          0.429
#>  8 bst-procontrol-gmbh_00000005104947-001 tilt_sec   medium        0.286
#>  9 bst-procontrol-gmbh_00000005104947-001 tilt_sec   low           0.286
#> 10 bst-procontrol-gmbh_00000005104947-001 unit       high          0.286
#> # ℹ 134 more rows

# Same
ictr(companies, co2)
#> # A tibble: 144 × 4
#>    companies_id                           grouped_by risk_category value
#>    <chr>                                  <chr>      <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 all        high          0.143
#>  2 bst-procontrol-gmbh_00000005104947-001 all        medium        0.429
#>  3 bst-procontrol-gmbh_00000005104947-001 all        low           0.429
#>  4 bst-procontrol-gmbh_00000005104947-001 isic_sec   high          0.429
#>  5 bst-procontrol-gmbh_00000005104947-001 isic_sec   medium        0.286
#>  6 bst-procontrol-gmbh_00000005104947-001 isic_sec   low           0.286
#>  7 bst-procontrol-gmbh_00000005104947-001 tilt_sec   high          0.429
#>  8 bst-procontrol-gmbh_00000005104947-001 tilt_sec   medium        0.286
#>  9 bst-procontrol-gmbh_00000005104947-001 tilt_sec   low           0.286
#> 10 bst-procontrol-gmbh_00000005104947-001 unit       high          0.286
#> # ℹ 134 more rows