Calculate the ISTR indicator

The Input Product Sector Risk Indicator assesses the transition risk of the input products based on the sector's emissions targets the input product belongs to. Those sector emission reduction targets vary across scenarios (e.g., net zero scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040, 2050 to achieve the targets). It, therefore, is similar to the Product Sector Risk Indicator but focuses on the input products that the company needs to produce its products.The input products are, for example, resources, packaging materials, energy and enabling services (such as tractor use on farm) to produce the product.

After identifying each carbon footprint for one input product, the input products are ranked according to their footprint. The ranking method is explained in the Thresholds section.

After assessing the input products for each product, they are aggregated at company-level to derive what percentage of the input products required by the company to produce its products have high, medium and low sector transition risk. We, therefore, derive the company-level information.

Please note that carbon emissions or emissions always mean CO2e.

Usage

istr(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

istr_at_product_level(
  companies,
  scenarios,
  inputs,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

istr_at_company_level(data)

Arguments

companies

A dataframe like istr_companies.

scenarios

A dataframe like xstr_scenarios.

inputs

A dataframe like istr_inputs.

low_threshold

A numeric value to segment low and medium transition risk products.

high_threshold

A numeric value to segment medium and high transition risk products.

data

A dataframe. The output at product level.

Value

A data frame with the column companies_id, and the nested columns product and company holding the outputs at product and company level. Unnesting product yields a data frame with at least columns companies_id, grouped_by, risk_category. Unnesting company yields a data frame with at least columns companies_id, grouped_by, risk_category, value.

Examples

companies <- istr_companies
scenarios <- xstr_scenarios
inputs <- istr_inputs

# Product level
companies |>
  istr_at_product_level(scenarios, inputs)
#> # A tibble: 704 × 12
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  2 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  3 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  4 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  5 fleischerei-stiefs… weo_annou… low           stove     0a242b09-772a-5edf-8e…
#>  6 fleischerei-stiefs… weo_annou… high          stove     0a242b09-772a-5edf-8e…
#>  7 fleischerei-stiefs… weo_annou… high          stove     0a242b09-772a-5edf-8e…
#>  8 fleischerei-stiefs… weo_annou… high          stove     0a242b09-772a-5edf-8e…
#>  9 fleischerei-stiefs… weo_net z… low           stove     0a242b09-772a-5edf-8e…
#> 10 fleischerei-stiefs… weo_net z… high          stove     0a242b09-772a-5edf-8e…
#> # ℹ 694 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 7 more variables: tilt_sector <chr>, scenario <chr>, year <dbl>,
#> #   type <chr>, input_activity_uuid_product_uuid <chr>,
#> #   input_tilt_sector <chr>, input_tilt_subsector <chr>

# Company level
companies |>
  istr_at_product_level(scenarios, inputs) |>
  istr_at_company_level()
#> # A tibble: 294 × 4
#>    companies_id                           grouped_by         risk_category value
#>    <chr>                                  <chr>              <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2030  high            0.2
#>  2 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2030  medium          0.2
#>  3 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2030  low             0.6
#>  4 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2050  high            1  
#>  5 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2050  medium          0  
#>  6 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2050  low             0  
#>  7 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… high            0  
#>  8 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… medium          0  
#>  9 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… low             1  
#> 10 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… high            0  
#> # ℹ 284 more rows

# Same
both <- istr(companies, scenarios, inputs)
both
#> # A tibble: 7 × 3
#>   companies_id                             product             company          
#>   <chr>                                    <list>              <list>           
#> 1 fleischerei-stiefsohn_00000005219477-001 <tibble [180 × 11]> <tibble [42 × 3]>
#> 2 pecheries-basques_fra316541-00101        <tibble [14 × 11]>  <tibble [42 × 3]>
#> 3 hoche-butter-gmbh_deu422723-693847001    <tibble [70 × 11]>  <tibble [42 × 3]>
#> 4 vicquelin-espaces-verts_fra697272-00101  <tibble [70 × 11]>  <tibble [42 × 3]>
#> 5 bst-procontrol-gmbh_00000005104947-001   <tibble [70 × 11]>  <tibble [42 × 3]>
#> 6 leider-gmbh_00000005064318-001           <tibble [150 × 11]> <tibble [42 × 3]>
#> 7 cheries-baqu_neu316541-00101             <tibble [150 × 11]> <tibble [42 × 3]>

both |> unnest_product()
#> # A tibble: 704 × 12
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  2 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  3 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  4 fleischerei-stiefs… weo_state… low           stove     0a242b09-772a-5edf-8e…
#>  5 fleischerei-stiefs… weo_annou… low           stove     0a242b09-772a-5edf-8e…
#>  6 fleischerei-stiefs… weo_annou… high          stove     0a242b09-772a-5edf-8e…
#>  7 fleischerei-stiefs… weo_annou… high          stove     0a242b09-772a-5edf-8e…
#>  8 fleischerei-stiefs… weo_annou… high          stove     0a242b09-772a-5edf-8e…
#>  9 fleischerei-stiefs… weo_net z… low           stove     0a242b09-772a-5edf-8e…
#> 10 fleischerei-stiefs… weo_net z… high          stove     0a242b09-772a-5edf-8e…
#> # ℹ 694 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 7 more variables: tilt_sector <chr>, scenario <chr>, year <dbl>,
#> #   type <chr>, input_activity_uuid_product_uuid <chr>,
#> #   input_tilt_sector <chr>, input_tilt_subsector <chr>

both |> unnest_company()
#> # A tibble: 294 × 4
#>    companies_id                             grouped_by      risk_category  value
#>    <chr>                                    <chr>           <chr>          <dbl>
#>  1 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_2… high          0.333 
#>  2 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_2… medium        0.583 
#>  3 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_2… low           0.0833
#>  4 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_2… high          1     
#>  5 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_2… medium        0     
#>  6 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_2… low           0     
#>  7 fleischerei-stiefsohn_00000005219477-001 weo_announced … high          0     
#>  8 fleischerei-stiefsohn_00000005219477-001 weo_announced … medium        0     
#>  9 fleischerei-stiefsohn_00000005219477-001 weo_announced … low           1     
#> 10 fleischerei-stiefsohn_00000005219477-001 weo_announced … high          0.0769
#> # ℹ 284 more rows