Calculate the PSTR indicator

The Product Sector Transition Risk Indicator measures the transition risk of products based on the sector's emissions targets the product belongs to. Those sector emission reduction targets vary across scenarios (e.g., net zero scenario or 1.5° scenario) and the time horizon (e.g., reduction needed in 2030, 2040, 2050 to achieve the targets).

After assessing each product, all the products with the same category are aggregated and set in relation to all products of the company. We, therefore, derive company-level information.

Usage

pstr(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_product_level(
  companies,
  scenarios,
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

pstr_at_company_level(data)

Arguments

companies

A dataframe like pstr_companies.

scenarios

A dataframe like xstr_scenarios.

low_threshold

A numeric value to segment low and medium reduction targets.

high_threshold

A numeric value to segment medium and high reduction targets.

data

A dataframe. The output at product level.

Value

At product level, a dataframe with at least columns companies_id, grouped_by, risk_category. At company level, a dataframe with at least columns companies_id, grouped_by, risk_category, value.

Examples

library(dplyr, warn.conflicts = FALSE)
library(tidyr)

companies <- pstr_companies
scenarios <- xstr_scenarios

# Product level
companies |>
  pstr_at_product_level(scenarios)
#> # A tibble: 196 × 10
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… ipr_1.5c … high          steel     0faa7ecb-fef2-5117-89…
#>  2 fleischerei-stiefs… ipr_1.5c … high          steel     0faa7ecb-fef2-5117-89…
#>  3 fleischerei-stiefs… weo_state… low           steel     0faa7ecb-fef2-5117-89…
#>  4 fleischerei-stiefs… weo_annou… low           steel     0faa7ecb-fef2-5117-89…
#>  5 fleischerei-stiefs… weo_net z… low           steel     0faa7ecb-fef2-5117-89…
#>  6 fleischerei-stiefs… weo_state… low           steel     0faa7ecb-fef2-5117-89…
#>  7 fleischerei-stiefs… weo_annou… low           steel     0faa7ecb-fef2-5117-89…
#>  8 fleischerei-stiefs… weo_net z… high          steel     0faa7ecb-fef2-5117-89…
#>  9 fleischerei-stiefs… weo_state… low           steel     0faa7ecb-fef2-5117-89…
#> 10 fleischerei-stiefs… weo_annou… medium        steel     0faa7ecb-fef2-5117-89…
#> # ℹ 186 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 5 more variables: tilt_sector <chr>, scenario <chr>, year <dbl>,
#> #   type <chr>, tilt_subsector <chr>

# Company level
companies |>
  pstr_at_product_level(scenarios) |>
  pstr_at_company_level()
#> # A tibble: 588 × 4
#>    companies_id                           grouped_by         risk_category value
#>    <chr>                                  <chr>              <chr>         <dbl>
#>  1 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2030  high              0
#>  2 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2030  medium            0
#>  3 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2030  low               1
#>  4 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2050  high              1
#>  5 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2050  medium            0
#>  6 bst-procontrol-gmbh_00000005104947-001 ipr_1.5c rps_2050  low               0
#>  7 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… high              0
#>  8 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… medium            0
#>  9 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… low               1
#> 10 bst-procontrol-gmbh_00000005104947-001 weo_announced ple… high              0
#> # ℹ 578 more rows

# Or
both <- pstr(companies, scenarios)
both
#> # A tibble: 14 × 3
#>    companies_id                             product           company          
#>    <chr>                                    <list>            <list>           
#>  1 fleischerei-stiefsohn_00000005219477-001 <tibble [14 × 9]> <tibble [42 × 3]>
#>  2 pecheries-basques_fra316541-00101        <tibble [14 × 9]> <tibble [42 × 3]>
#>  3 hoche-butter-gmbh_deu422723-693847001    <tibble [14 × 9]> <tibble [42 × 3]>
#>  4 hoche-butter-gmbh_deu422723-693847002    <tibble [14 × 9]> <tibble [42 × 3]>
#>  5 hoche-butter-gmbh_deu422723-693847003    <tibble [14 × 9]> <tibble [42 × 3]>
#>  6 vicquelin-espaces-verts_fra697272-00101  <tibble [14 × 9]> <tibble [42 × 3]>
#>  7 vicquelin-espaces-verts_fra697272-00102  <tibble [14 × 9]> <tibble [42 × 3]>
#>  8 vicquelin-espaces-verts_fra697272-00103  <tibble [14 × 9]> <tibble [42 × 3]>
#>  9 fleisohn_0000000492-001                  <tibble [14 × 9]> <tibble [42 × 3]>
#> 10 bst-procontrol-gmbh_00000005104947-001   <tibble [14 × 9]> <tibble [42 × 3]>
#> 11 leider-gmbh_00000005064318-001           <tibble [14 × 9]> <tibble [42 × 3]>
#> 12 leider-gmbh_00000005064318-002           <tibble [14 × 9]> <tibble [42 × 3]>
#> 13 cheries-baqu_neu316541-00101             <tibble [14 × 9]> <tibble [42 × 3]>
#> 14 ca-coity-trg-aua-gmbh_00000384-001       <tibble [14 × 9]> <tibble [42 × 3]>

# Product level
both |> unnest_product()
#> # A tibble: 196 × 10
#>    companies_id        grouped_by risk_category clustered activity_uuid_produc…¹
#>    <chr>               <chr>      <chr>         <chr>     <chr>                 
#>  1 fleischerei-stiefs… ipr_1.5c … high          steel     0faa7ecb-fef2-5117-89…
#>  2 fleischerei-stiefs… ipr_1.5c … high          steel     0faa7ecb-fef2-5117-89…
#>  3 fleischerei-stiefs… weo_state… low           steel     0faa7ecb-fef2-5117-89…
#>  4 fleischerei-stiefs… weo_annou… low           steel     0faa7ecb-fef2-5117-89…
#>  5 fleischerei-stiefs… weo_net z… low           steel     0faa7ecb-fef2-5117-89…
#>  6 fleischerei-stiefs… weo_state… low           steel     0faa7ecb-fef2-5117-89…
#>  7 fleischerei-stiefs… weo_annou… low           steel     0faa7ecb-fef2-5117-89…
#>  8 fleischerei-stiefs… weo_net z… high          steel     0faa7ecb-fef2-5117-89…
#>  9 fleischerei-stiefs… weo_state… low           steel     0faa7ecb-fef2-5117-89…
#> 10 fleischerei-stiefs… weo_annou… medium        steel     0faa7ecb-fef2-5117-89…
#> # ℹ 186 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 5 more variables: tilt_sector <chr>, scenario <chr>, year <dbl>,
#> #   type <chr>, tilt_subsector <chr>

# Company level
both |> unnest_company()
#> # A tibble: 588 × 4
#>    companies_id                             grouped_by       risk_category value
#>    <chr>                                    <chr>            <chr>         <dbl>
#>  1 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_20… high              1
#>  2 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_20… medium            0
#>  3 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_20… low               0
#>  4 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_20… high              1
#>  5 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_20… medium            0
#>  6 fleischerei-stiefsohn_00000005219477-001 ipr_1.5c rps_20… low               0
#>  7 fleischerei-stiefsohn_00000005219477-001 weo_announced p… high              0
#>  8 fleischerei-stiefsohn_00000005219477-001 weo_announced p… medium            0
#>  9 fleischerei-stiefsohn_00000005219477-001 weo_announced p… low               1
#> 10 fleischerei-stiefsohn_00000005219477-001 weo_announced p… high              0
#> # ℹ 578 more rows